src/backend/face_recognition_app/requirements.txt
Flask==0.12.2
faiss-cpu==1.6.3
flask-restplus==0.13.0
pymilvus==0.2.13
face-recognition==1.3.0
mtcnn==0.1.0
opencv-python==4.2.0.34
Werkzeug==0.16.0
tensorflow==2.2.0
boto3==1.14.13
face-alignment==1.0.0
